Restructure GraphQL automated content scripts (#34308)
