+++ src/graphql/README.md
@@ -2,29 +2,27 @@
 
 ## About this directory
 
-* `lib/graphql/*.json` are human-editable.
-* `lib/graphql/static/*.json` are generated by [scripts](../script/graphql/README.md).
+* `src/graphql/lib` and `src/graphql/scripts` are human-editable.
+* `src/graphql/data/**` are generated by [scripts](../src/graphql/README.md).
 
 ## Editable files
 
-* `lib/graphql/validator.json`
+* `src/graphql/lib/validator.json`
   - JSON schema used in `tests/graphql.js`.
-* `lib/graphql/non-schema-scalars.json`
+* `src/graphql/lib/non-schema-scalars.json`
   - An array of scalar types that live in [`graphql-ruby`](https://github.com/rmosolgo/graphql-ruby/tree/356d9d369e444423bf06cab3dc767ec75fbc6745/lib/graphql/types) only. These are
   not part of the core GraphQL spec.
-* `lib/graphql/redirects.json`
-  - A list of original `developer.github.com/v4` paths and their corresponding new `docs.github.com/graphql` paths. Used in redirect middleware and tests.
-* `lib/graphql/types.json`
+* `src/graphql/lib/types.json`
   - High-level GraphQL types and kinds.
 
-## Static files
+## Data files
 
-Generated by `script/graphql/update-files.js`:
+Generated by `src/graphql/scripts/update-files.js`:
 
-* `lib/graphql/static/schema-VERSION.json` (separate files per version)
-* `lib/graphql/static/previews.json`
-* `lib/graphql/static/upcoming-changes.json`
-* `lib/graphql/static/changelog.json`
+* `src/graphql/data/schema-VERSION.json` (separate files per version)
+* `src/graphql/data/previews.json`
+* `src/graphql/data/upcoming-changes.json`
+* `src/graphql/data/changelog.json`
 
 ## Rendering docs
