+++ src/graphql/data/schema-ghes-3.6.json
