+++ src/graphql/scripts/utils/process-previews.js
