+++ src/graphql/scripts/utils/schema-helpers.js
@@ -1,11 +1,11 @@
 #!/usr/bin/env node
-import renderContent from '../../../lib/render-content/index.js'
+import renderContent from '../../../../lib/render-content/index.js'
 import fs from 'fs/promises'
 import graphql from 'graphql'
 import path from 'path'
 
 const graphqlTypes = JSON.parse(
-  await fs.readFile(path.join(process.cwd(), './lib/graphql/types.json'))
+  await fs.readFile(path.join(process.cwd(), './src/graphql/lib/types.json'))
 )
 const { isScalarType, isObjectType, isInterfaceType, isUnionType, isEnumType, isInputObjectType } =
   graphql
